Frontend: Move regex frontend tests into respective test module
